% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AREAOFINFERENCES}
\alias{AREAOFINFERENCES}
\title{AREAOFINFERENCES}
\format{
\subsection{\code{AREAOFINFERENCES}}{

data frame created from AREAOFINFERENCES.csv
\describe{
\item{Data Use}{NA}
}
}
}
\source{
\url{https://cbfish.org/}
}
\usage{
AREAOFINFERENCES
}
\description{
CBFish Area of Inference
}
\keyword{datasets}
