% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMETRICS}
\alias{EMETRICS}
\title{EMETRICS}
\format{
\subsection{\code{EMETRICS}}{

data frame created from PORTFOLIOENVIRONMENTALMETRICS.csv
\describe{
\item{Data Use}{NA}
}
}
}
\source{
\url{https://cbfish.org/}
}
\usage{
EMETRICS
}
\description{
Environmental Metrics
}
\keyword{datasets}
