% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLSpendingSites.R
\name{HTML.getWorksiteFiles}
\alias{HTML.getWorksiteFiles}
\title{Returns unique worksites for Work Site ID, Description, and Lat Lon
Save result and import into QGIS to get HUC_12 attribute
Also returns data frame connecting WSEs to Work Site ID}
\usage{
HTML.getWorksiteFiles(
  worksites = CBFishR.BPA::WORKSITES,
  sf_polygons = CBFishR.BPA::HUC12Polys,
  hucs_pops = CBFishR.BPA::HUCsPOPs
)
}
\arguments{
\item{worksites}{data frame}
}
\value{
list of two data frames
\describe{
\item{Locs}{NA}
\item{WSE_Sites}{NA}
}
}
\description{
Returns unique worksites for Work Site ID, Description, and Lat Lon
Save result and import into QGIS to get HUC_12 attribute
Also returns data frame connecting WSEs to Work Site ID
}
\examples{
HTML.getWorksiteFiles()

}
