% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMLSpendingSites.R
\name{HTML.getWSEFile}
\alias{HTML.getWSEFile}
\title{Process WSE file to convert dollars and percents to numbers and such}
\usage{
HTML.getWSEFile(
  projects = CBFishR.BPA::PROJECTS,
  contracts = CBFishR.BPA::CONTRACTS,
  we = CBFishR.BPA::WORKELEMENTS,
  wse
)
}
\value{
list of two data frames
\describe{
\item{NA}{NA}
}
}
\description{
Process WSE file to convert dollars and percents to numbers and such
}
\examples{
HTML.getWSEFile()

}
