% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HUC12Polys}
\alias{HUC12Polys}
\title{HUC12Polys}
\format{
\subsection{\code{HUCsPOPs}}{

sf object
\describe{
\item{Data Use}{NA}
}
}
}
\source{
\url{https://cbfish.org/}
}
\usage{
HUC12Polys
}
\description{
HUC12 shapefile using package sf
}
\keyword{datasets}
