% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HUCsPOPs}
\alias{HUCsPOPs}
\title{HUCsPOPs}
\format{
\subsection{\code{HUCsPOPs}}{

data frame created from raw-data/HUCs_POPs.csv
\describe{
\item{Data Use}{NA}
}
}
}
\source{
\url{https://cbfish.org/}
}
\usage{
HUCsPOPs
}
\description{
Mapping of populations to HUC12
}
\keyword{datasets}
