% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMETRICS.NUMERIC}
\alias{IMETRICS.NUMERIC}
\title{IMETRICS.NUMERIC}
\format{
\subsection{\code{IMETRICS.NUMERIC}}{

data frame created by filtering
\describe{
\item{Data Use}{NA}
}
}
}
\source{
\url{https://cbfish.org/}
}
\usage{
IMETRICS.NUMERIC
}
\description{
Numerical Implementation Metrics
}
\keyword{datasets}
