% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMETRICS}
\alias{IMETRICS}
\title{IMETRICS}
\format{
\subsection{\code{IMETRICS}}{

data frame created from PORTFOLIOIMPLEMENTATIONMETRICS.csv
\describe{
\item{Data Use}{NA}
}
}
}
\source{
\url{https://cbfish.org/}
}
\usage{
IMETRICS
}
\description{
Implementation Metrics
}
\keyword{datasets}
