% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEASURES}
\alias{MEASURES}
\title{MEASURES}
\format{
\subsection{\code{MEASURES}}{

data frame created from MEASURES.csv
\describe{
\item{Data Use}{NA}
}
}
}
\source{
\url{https://cbfish.org/}
}
\usage{
MEASURES
}
\description{
CBFish Measures
}
\keyword{datasets}
