% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{METRICS}
\alias{METRICS}
\title{METRICS}
\format{
\subsection{\code{METRICS}}{

data frame created from METRICS.csv
\describe{
\item{Data Use}{NA}
}
}
}
\source{
\url{https://cbfish.org/}
}
\usage{
METRICS
}
\description{
CBFish Metrics
}
\keyword{datasets}
