% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROJECTS}
\alias{PROJECTS}
\title{PROJECTS}
\format{
\subsection{\code{PROJECTS}}{

data frame PORTFOLIO-AllProjects-Performance-projects.csv
\describe{
\item{Data Use}{NA}
}
}
}
\source{
\url{https://cbfish.org/}
}
\usage{
PROJECTS
}
\description{
CBFish Projects
}
\keyword{datasets}
