% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setUnspecified}
\alias{setUnspecified}
\title{Set values="Unspecified" if ""}
\usage{
setUnspecified(name)
}
\arguments{
\item{name}{String vector}
}
\value{
String vector
}
\description{
Set values="Unspecified" if ""
}
\examples{
setUnspecified(example)

}
