% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trimMult}
\alias{trimMult}
\title{I have no idea}
\usage{
trimMult(x, char = " ")
}
\arguments{
\item{amt}{String vector}
}
\value{
String vector
}
\description{
I have no idea
}
\examples{
trimMult(example)

}
