% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unMoney}
\alias{unMoney}
\title{Remove the dollar sign and commas and converts numbers within parentheses to negative numbers}
\usage{
unMoney(monetaryStrings)
}
\arguments{
\item{monetaryStrings}{String vector}
}
\value{
numeric vector
}
\description{
Remove the dollar sign and commas and converts numbers within parentheses to negative numbers
}
\examples{
unMoney(moneystr)

}
