% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unPct}
\alias{unPct}
\title{Remove the percent sign and divides by 100}
\usage{
unPct(amt)
}
\arguments{
\item{amt}{String vector}
}
\value{
numeric vector
}
\description{
Remove the percent sign and divides by 100
}
\examples{
unPct(pctstr)

}
