% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WORKELEMENTS}
\alias{WORKELEMENTS}
\title{WORKELEMENTS}
\format{
\subsection{\code{WORKELEMENTS}}{

data frame created from WORKELEMENTS.csv
\describe{
\item{Data Use}{NA}
}
}
}
\source{
\url{https://cbfish.org/}
}
\usage{
WORKELEMENTS
}
\description{
CBFish Work Elements
}
\keyword{datasets}
