% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WORKSITES}
\alias{WORKSITES}
\title{WORKSITES}
\format{
\subsection{\code{WORKSITES}}{

data frame WORKSITES.csv
\describe{
\item{Data Use}{NA}
}
}
}
\source{
\url{https://cbfish.org/}
}
\usage{
WORKSITES
}
\description{
Work Sites
}
\keyword{datasets}
