% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WSE}
\alias{WSE}
\title{WSE}
\format{
\subsection{\code{WSE}}{

data frame PORTFOLIO-AllProjects-Performance-WorkStatementElements.csv
\describe{
\item{Data Use}{NA}
}
}
}
\source{
\url{https://cbfish.org/}
}
\usage{
WSE
}
\description{
Work Statement Elements
}
\keyword{datasets}
